requirements.txt
# Install pydantic-ai with MCP and Groq support
pydantic-ai[mcp,groq]==0.1.3

# MCP fetch server package
mcp-server-fetch==2025.4.7

# Manage environment variables
python-dotenv

# Streamlit for frontend
streamlit

# URL validation
validators

# HTTP requests
requests